New/updated tests for default connections and drp
I think the new test here is `nw-drp-001` which checks that the default
connection for a secondary input port is its default connection. I’m surprised
there wasn’t a test for this, but apparently there wasn’t.

I’ve also updated the allowed error codes for several drp-related tests so that
either err:XS0003 or err:XS0032 is allowed.