Adds new p:os-exec tests
These tests use a little Java program that can be instructed to produce output
in various ways. (I’ll get the sources checked in too, asap.)

The first test just makes sure it works. The second test demonstrates that
without a charset option, the output is wrong. The third test demonstrates that
a charset option can correct the output problem. (But of course, will fail if
the implementation doesn’t support that not-yet-standard option.)