Reworked tests to produce a single result document
Schematron can handle a document that contains more than one top-level element,
but the result is difficult to reparse. Reparsing the test results was useful in
finding a couple of namespace-related errors in my implementation, so I’ve
reworked the few tests that produce sequences so that they produce a single
result.

The changes are all trivial except for ab-xslt-062 where I think I’ve improved
the test.